Add function
Execute a touch event based on the absolute coordinates and complete a function that reads the sentence existing at the location.
After moving the calibration-related variables to the setting activity, the environment is being built again.
