+++ EyePedia/app/src/main/java/com/example/eyepedia/SettingActivity.java
@@ -3,10 +3,28 @@ package com.example.eyepedia;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+import android.os.Handler;
+import android.os.HandlerThread;
 import android.util.Log;
+import android.view.View;
+import android.widget.Button;
+import android.widget.Toast;
+
+import com.example.eyepedia.view.CalibrationDataStorage;
+import com.example.eyepedia.view.CalibrationViewer;
+import com.example.eyepedia.view.PointView;
+
+import camp.visual.gazetracker.GazeTracker;
+import camp.visual.gazetracker.callback.CalibrationCallback;
+import camp.visual.gazetracker.constant.CalibrationModeType;
+import camp.visual.gazetracker.util.ViewLayoutChecker;
 
 public class SettingActivity extends AppCompatActivity {
     private static final String TAG = SettingActivity.class.getSimpleName();
+    private GazeTracker gazeTracker;
+    private ViewLayoutChecker viewLayoutChecker = new ViewLayoutChecker();
+    private HandlerThread backgroundThread = new HandlerThread("background");
+    private Handler backgroundHandler;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -15,4 +33,199 @@ public class SettingActivity extends AppCompatActivity {
 
         Log.i(TAG, "Setiing open Completed");
     }
+
+    // handler
+    private void initHandler() {
+        backgroundThread.start();
+        backgroundHandler = new Handler(backgroundThread.getLooper());
+    }
+
+    private void releaseHandler() {
+        backgroundThread.quitSafely();
+    }
+    // handler end
+
+    //view
+    private Button btnStartCalibration, btnStopCalibration, btnSetCalibration;
+    private PointView viewPoint;
+    private CalibrationViewer viewCalibration;
+    private CalibrationModeType calibrationType = CalibrationModeType.DEFAULT;
+
+    private void initView() {
+        btnStartCalibration = findViewById(R.id.btn_start_calibration);
+        btnStopCalibration = findViewById(R.id.btn_stop_calibration);
+        btnStartCalibration.setOnClickListener(onClickListener);
+        btnStopCalibration.setOnClickListener(onClickListener);
+        btnSetCalibration = findViewById(R.id.btn_set_calibration);
+        btnSetCalibration.setOnClickListener(onClickListener);
+
+        viewPoint = findViewById(R.id.view_point);
+        viewCalibration = findViewById(R.id.view_calibration);
+
+        setOffsetOfView();
+    }
+
+    private void setOffsetOfView() {
+        viewLayoutChecker.setOverlayView(viewPoint, new ViewLayoutChecker.ViewLayoutListener() {
+            @Override
+            public void getOffset(int x, int y) {
+                viewCalibration.setOffset(x, y);
+            }
+        });
+    }
+
+    private View.OnClickListener onClickListener = new View.OnClickListener() {
+        @Override
+        public void onClick(View v) {
+            if (v == btnStartCalibration) {
+                startCalibration();
+            } else if (v == btnStopCalibration) {
+                stopCalibration();
+            } else if (v == btnSetCalibration) {
+                setCalibration();
+            }
+        }
+    };
+
+    private boolean startCalibration() {
+        boolean isSuccess = false;
+        if (isGazeNonNull()) {
+            isSuccess = gazeTracker.startCalibration(calibrationType);
+            if (!isSuccess) {
+                showToast("calibration start fail", false);
+            }
+        }
+        setViewAtGazeTrackerState();
+        return isSuccess;
+    }
+
+    // Collect the data samples used for calibration
+    private boolean startCollectSamples() {
+        boolean isSuccess = false;
+        if (isGazeNonNull()) {
+            isSuccess = gazeTracker.startCollectSamples();
+        }
+        setViewAtGazeTrackerState();
+        return isSuccess;
+    }
+
+    private void stopCalibration() {
+        if (isGazeNonNull()) {
+            gazeTracker.stopCalibration();
+        }
+        hideCalibrationView();
+        setViewAtGazeTrackerState();
+    }
+
+    private void setCalibration() {
+        if (isGazeNonNull()) {
+            double[] calibrationData = CalibrationDataStorage.loadCalibrationData(getApplicationContext());
+            if (calibrationData != null) {
+                // When if stored calibration data in SharedPreference
+                if (!gazeTracker.setCalibrationData(calibrationData)) {
+                    showToast("calibrating", false);
+                } else {
+                    showToast("setCalibrationData success", false);
+                }
+            } else {
+                // When if not stored calibration data in SharedPreference
+                showToast("Calibration data is null", true);
+            }
+        }
+        setViewAtGazeTrackerState();
+    }
+
+    private void showToast(final String msg, final boolean isShort) {
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                Toast.makeText(SettingActivity.this, msg, isShort ? Toast.LENGTH_SHORT : Toast.LENGTH_LONG).show();
+            }
+        });
+    }
+
+    private void setCalibrationPoint(final float x, final float y) {
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                viewCalibration.setVisibility(View.VISIBLE);
+//                menuBar.setVisibility(View.INVISIBLE);
+                viewCalibration.changeDraw(true, null);
+                viewCalibration.setPointPosition(x, y);
+                viewCalibration.setPointAnimationPower(0);
+            }
+        });
+    }
+
+    private void setCalibrationProgress(final float progress) {
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                viewCalibration.setPointAnimationPower(progress);
+            }
+        });
+    }
+
+    private void hideCalibrationView() {
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                viewCalibration.setVisibility(View.INVISIBLE);
+//                menuBar.setVisibility(View.VISIBLE);
+            }
+        });
+    }
+
+    private void setViewAtGazeTrackerState() {
+//        Log.i(TAG, "gaze : " + isGazeNonNull() + ", tracking " + isTracking());
+        runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                if (!isTracking()) {
+                    hideCalibrationView();
+                }
+            }
+        });
+    }
+    // view end
+
+    // gazeTracker
+    private boolean isTracking() {
+        if (isGazeNonNull()) {
+            return gazeTracker.isTracking();
+        }
+        return false;
+    }
+
+    private boolean isGazeNonNull() {
+        return gazeTracker != null;
+    }
+
+    private CalibrationCallback calibrationCallback = new CalibrationCallback() {
+        @Override
+        public void onCalibrationProgress(float progress) {
+            setCalibrationProgress(progress);
+        }
+
+        @Override
+        public void onCalibrationNextPoint(final float x, final float y) {
+            setCalibrationPoint(x, y);
+            // Give time to eyes find calibration coordinates, then collect data samples
+            backgroundHandler.postDelayed(new Runnable() {
+                @Override
+                public void run() {
+                    startCollectSamples();
+                }
+            }, 1000);
+        }
+
+        @Override
+        public void onCalibrationFinished(double[] calibrationData) {
+            // 캘리브레이션이 끝나면 자동으로 gazepoint에 적용되어있고
+            // calibrationDataStorage에 calibrationData를 넣는것은 다음번에 캘리브레이션 하지않고 사용하게 하기 위함이다.
+            CalibrationDataStorage.saveCalibrationData(getApplicationContext(), calibrationData);
+            hideCalibrationView();
+            showToast("calibrationFinished", true);
+        }
+    };
 }
