+++ EyePedia/app/src/main/res/layout/activity_main.xml
@@ -22,56 +22,10 @@
     </com.google.android.material.appbar.AppBarLayout>
 
     <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/layout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_behavior="@string/appbar_scrolling_view_behavior">
-
-        <LinearLayout
-            android:id="@+id/linearLayout"
-            android:layout_width="200dp"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="24dp"
-            android:orientation="vertical"
-            android:visibility="visible"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintStart_toStartOf="parent">
-
-            <Button
-                android:id="@+id/btn_start_calibration"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="정확도 향상" />
-
-            <Button
-                android:id="@+id/btn_set_calibration"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="설정 저장" />
-
-            <Button
-                android:id="@+id/btn_stop_calibration"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="설정 삭제" />
-
-            <Button
-                android:text="파일읽기"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="mOnFileRead"/>
-
-            <TextView
-                android:id="@+id/txtRead"
-                android:layout_width="match_parent"
-                android:layout_height="51dp"
-                android:text="TextView"
-                tools:layout_editor_absoluteX="179dp"
-                tools:layout_editor_absoluteY="153dp" />
-
-
-        </LinearLayout>
 
         <TextView
             android:id="@+id/textView"
@@ -87,19 +41,41 @@
             app:layout_constraintHorizontal_bias="0.0"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="0.0" />
+            app:layout_constraintVertical_bias="1.0" />
+
+        <LinearLayout
+            android:id="@+id/linearLayout"
+            android:layout_width="250dp"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="24dp"
+            android:orientation="vertical"
+            android:visibility="visible"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintRight_toRightOf="parent"
+            app:layout_constraintStart_toStartOf="parent">
+
+            <Button
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:onClick="mOnFileRead"
+                android:text="파일읽기" />
+
+            <Button
+                android:id="@+id/button"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="Button"
+                android:textSize="20sp" />
+        </LinearLayout>
+
+
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 
     <!-- 시선 좌표 -->
     <com.example.eyepedia.view.PointView
         android:id="@+id/view_point"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-    <!-- 캘리브레이션 뷰 -->
-    <com.example.eyepedia.view.CalibrationViewer
-        android:id="@+id/view_calibration"
-        android:visibility="invisible"
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
